+++ apps/probability/models/statistic/two_proportions_statistic.h
@@ -8,9 +8,9 @@ namespace Probability {
 class TwoProportionsStatistic : public ZStatistic {
 public:
   void init(Data::SubApp subapp) override;
-  bool isValidH0(float p) override;
-  bool isValidParamAtIndex(int i, float p) override;
-  void setParamAtIndex(int index, float p) override;
+  bool isValidH0(double p) override;
+  bool isValidParamAtIndex(int i, double p) override;
+  void setParamAtIndex(int index, double p) override;
 
   void computeTest() override;
   void computeInterval() override;
@@ -29,22 +29,22 @@ protected:
   int numberOfStatisticParameters() const override { return k_numberOfParams; }
   enum ParamsOrder { X1, N1, X2, N2 };
   ParameterRepresentation paramRepresentationAtIndex(int i) const override;
-  float * paramArray() override { return m_params; }
+  double * paramArray() override { return m_params; }
 
   bool validateInputs() override;
 
 private:
-  float x1() { return m_params[ParamsOrder::X1]; }
-  float n1() { return m_params[ParamsOrder::N1]; }
-  float x2() { return m_params[ParamsOrder::X2]; }
-  float n2() { return m_params[ParamsOrder::N2]; }
+  double x1() { return m_params[ParamsOrder::X1]; }
+  double n1() { return m_params[ParamsOrder::N1]; }
+  double x2() { return m_params[ParamsOrder::X2]; }
+  double n2() { return m_params[ParamsOrder::N2]; }
 
   // Computation
-  float computeEstimate(float x1, float n1, float x2, float n2);
-  float computeZ(float deltaP0, float x1, int n1, float x2, int n2);
-  float computeStandardError(float p1Estimate, int n1, float p2Estimate, int n2);
+  double computeEstimate(double x1, double n1, double x2, double n2);
+  double computeZ(double deltaP0, double x1, int n1, double x2, int n2);
+  double computeStandardError(double p1Estimate, int n1, double p2Estimate, int n2);
 
-  float m_params[k_numberOfParams];
+  double m_params[k_numberOfParams];
 };
 
 }  // namespace Probability
